requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
asyncpg
greenlet
pydantic
passlib[bcrypt]
python-jose[cryptography]
python-multipart
aiosqlite
httpx
